(svn r18016) -Codechange: Move the arbitrary basecost multiplier offset (8) to newgrf loading and make the internal state zero-based instead.

+++ src/economy_func.h
@@ -23,7 +23,7 @@
 #include "station_type.h"
 
 void ResetPriceBaseMultipliers();
-void SetPriceBaseMultiplier(Price price, byte factor);
+void SetPriceBaseMultiplier(Price price, int factor);
 
 extern const ScoreInfo _score_info[];
 extern int _score_part[MAX_COMPANIES][SCORE_END];
